Refactor maker graph and home hooks
